Renamed some util files
