+++ python/smqtk/utils/signal_handler.py
