+++ python/smqtk/utils/simple_timer.py
