+++ python/smqtk/utils/videofile.py
